Stable Version for two lenda Bars and the NA-22 source
Builds evens for two lenda bars and the NA-22 source.  Takes raw tree
from the ddasdumper as input and looks for groups of 4 that meet the following
conditions:
There are 4 distinict channels and none of those channels are 8 or 9 (the liquid scints)
The time difference between is less then 10 clock ticks
The time difference is defined as:
 timeDiff = 0.5*(events[0]->time + events[1]->time - events[2]->time-events[3]->time);

Updated the file management and input manangement.  Doesn't support the timing mode
specification anymore.  Now it has the notes: argument which will add whatever you put there
into the filename.  Defualts to 4Channel-Gamme in oder to distiunguish it from other modes

Still has the trace fitting and software filter inputs and filemanagment however it is NOT
saving the software time.  Right now only internal time is supported.