Did some tweaks and fixes.
Fixed packEvent again.  Had shortGate and longGate both being set with longGate input value.

Added startEntry feature to builder.  Allows you to set the start of the  main analysis loop
Added the necessary flag structure to Inputmanager.cc/.hh

Made the input and output file paths absolute paths instead of ./
This was done to allow the builder to be run on the seaside cluster.  The output and input lines now look
for ~/analysis/run{runNumber} structure.  Added a runNumber memeber variable to FileManager.